pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "lost"
version = "0.0.1"
description = "LoST Protocol Reimagined"
authors = [
  { name = "Jan Janak", email = "jan@janakj.org" }
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
  "flask",
  "click",
  "psycopg[binary]",
  "psycopg[pool]",
  "flask-cors",
  "flask-marshmallow",
  "pyproj",
  "lxml",
  "pygml",
  "requests",
  "osm2geojson",
  "tabulate",
  "scikit-image",
  "pillow"
]

[project.optional-dependencies]
dev = [
  "mypy",
  "types-flask-cors",
  "types-lxml",
  "types-requests"
]

[project.urls]
"Homepage" = "https://github.com/irtlab/lost"
"Bug Tracker" = "https://github.com/irtlab/lost/issues"

[project.scripts]
lost-server   = "lost.server:cli"
lost-resolver = "lost.resolver.cli:main"
lost-seeker   = "lost.seeker:cli"
lost-publisher   = "lost.publisher:cli"

[tool.setuptools]
packages = ["lost"]
